applied contets of pull request

+++ PIC_Timer0.c
@@ -9,7 +9,7 @@ void Timer0_Init(TMR0_PRESCALER_VAL pre_val,uint8_t reload_val)
     OPTION_REG &= ~(7<<0);  //Clear Prescaler value
     OPTION_REG |=  ((uint8_t)pre_val<<0);  //Set prescaler value
     
-    TMR0_ReloadTemp = 0xFF - reload_val;
+    TMR0_ReloadTemp = (uint8_t)(0xFF - reload_val);
     TMR0 = TMR0_ReloadTemp;   //Clear 
     
     TMR0IF = 0;
